+++ lexer/delimiters/string.c
@@ -1,7 +1,7 @@
 
 #include "../../minishell.h"
 
-void	fill_string_in_node(char *str, int *i, t_vars *data, char *delimiters)
+void	fill_string_in_node(char *str, size_t *i, t_vars *data, char *delimiters)
 {
 	Datatoken	*node;
 	char		*tmp;
